Move handle_month, parse_date_param, parse_number_of_days to validation.py
and fix imports, mocks
